Remove legacy code: remove CMAEvolutionStrategy.result() ordinary tuple fallback
`CMAEvolutionStrategy.result()` returns a CMAEvolutionStrategyResult namedtuple. However it contains a fallback code for cases where namedtuple is unavailable.

I believe this is obsolete legacy code: In `setup.py` the minimum targeted Python version is specified as 2.7. `namedtuple` was introduced in Python 2.6 so that this fallback should no longer be necessary.

Apart from code cleanup, a benefit of this change is that static type checkers can be confident about the return type of `CMAEvolutionStrategy.result()`. Currently static type checkers conclude that the return type may be `tuple` or `namedtuple`. As a result they can not be sure that a user may access the result as follows: `es.result().xfavorite` and will raise an error. (Since in cases where an ordinary tuple is returned the user code would be invalid.)

After this change, static type checkers can be confident about the return type being a named tuple and that attribute access is permissible.