Significant improvement in run times by addressing `SolutionDict#key()`
When scaling to large population sizes (e.g. IPOP), a significant portion of the wall-clock time for CMA-ES was taken by method `SolutionDict#key()`. This commit addresses the problem by (i) switching to the built-in hash function, (ii) skipping `np.ascontiguousarray` and the `data` attribute of NumPy arrays altogether, (iii) replacing the call to `self._hash` with a direct return statement, and (iv) choosing either a `tuple` or `bytes` transformation for Python 2 with an older NumPy version depending on the array size, overall leading to a significant reduction of run times. Moreover, new doctests were added that cover both parameters `x` of type `list` and `int`.